[cmake] Fix warnings showing up when client uses stronger warnings than library
Changes public include directories to be marked as system, so including them in a project with more warnings enabled won't fail.
Adds a copy of the includes as private and NOT system, to allow normal/CI builds to still have warnings.
Also changes other includes to be system, since they are system libraries.

Signed-off-by: Gabe Gonzalez II <lilggamegenius@gmail.com>